update: restructure to remove circular deps
