+++ src/py/mat3ra/made/tools/analyze/nanotape_analyzer.py
@@ -3,10 +3,8 @@ from typing import Optional
 from mat3ra.esse.models.core.reusable.axis_enum import AxisEnum
 from mat3ra.made.tools.analyze.base_stack_analyzer import BaseStackAnalyzer
 from mat3ra.made.tools.analyze.lattice_lines import CrystalLatticeLinesAnalyzer
-from mat3ra.made.tools.build.nanoribbon.configuration import (
-    CrystalLatticeLinesUniqueRepeatedConfiguration,
-    NanoTapeConfiguration,
-)
+from mat3ra.made.tools.build.nanoribbon.crystal_lattice_lines_configuration import CrystalLatticeLinesUniqueRepeatedConfiguration
+from mat3ra.made.tools.build.nanoribbon.nano_tape_configuration import NanoTapeConfiguration
 from mat3ra.made.tools.build.slab.entities import Termination
